requirements.txt
pytest
playwright
pytest-bdd
requests
mysql-connector-python  # Or psycopg2 for PostgreSQL, or sqlite3